Correction to the way of switching fragments.
Fixes #3 - Correction to how we handle fragment switching, check to see whether the fragment is within the backstack and if not add it to it. Otherwise we add it to the backstack.

Fixes issue relating to on back pressed main menu does not retain the name of the user within the textview.